Move formatDateRangeWithMessage to the StatusIndicator component
This puts it closer to where it's going to be used.
