+++ common/utils/format-date.ts
@@ -1,5 +1,3 @@
-import { isFuture, isPast, isSameDay } from './dates';
-
 export function formatLondon(
   date: Date,
   options: Intl.DateTimeFormatOptions
@@ -27,32 +25,6 @@ export function formatTime(date: Date): string {
   const minutes = formatLondon(date, { minute: '2-digit' }).padStart(2, '0');
 
   return `${hours}:${minutes}`;
-}
-
-export function formatDateRangeWithMessage({
-  start,
-  end,
-}: {
-  start: Date;
-  end: Date;
-}): { text: string; color: string } {
-  const today = new Date();
-
-  const sevenDaysTime = new Date();
-  sevenDaysTime.setDate(sevenDaysTime.getDate() + 7);
-
-  const closesToday = isSameDay(end, today);
-  const closesInSevenDays = today < end && end < sevenDaysTime;
-
-  if (!isSameDay(today, start) && isFuture(start)) {
-    return { text: 'Coming soon', color: 'marble' };
-  } else if (!isSameDay(today, end) && isPast(end)) {
-    return { text: 'Past', color: 'marble' };
-  } else if (closesToday || closesInSevenDays) {
-    return { text: 'Final week', color: 'orange' };
-  } else {
-    return { text: 'Now on', color: 'green' };
-  }
 }
 
 export function formatYear(date: Date): string {
