+++ content/webapp/components/StatusIndicator/StatusIndicator.tsx
@@ -1,14 +1,40 @@
 import { font } from '@weco/common/utils/classnames';
-import { formatDateRangeWithMessage } from '@weco/common/utils/format-date';
 import Space from '@weco/common/views/components/styled/Space';
 import Dot from '@weco/common/views/components/Dot/Dot';
 import { FC } from 'react';
+import { isFuture, isPast, isSameDay } from '@weco/common/utils/dates';
 
 type Props = {
   start: Date;
   end: Date;
   statusOverride?: string;
 };
+
+export function formatDateRangeWithMessage({
+  start,
+  end,
+}: {
+  start: Date;
+  end: Date;
+}): { text: string; color: string } {
+  const today = new Date();
+
+  const sevenDaysTime = new Date();
+  sevenDaysTime.setDate(sevenDaysTime.getDate() + 7);
+
+  const closesToday = isSameDay(end, today);
+  const closesInSevenDays = today < end && end < sevenDaysTime;
+
+  if (!isSameDay(today, start) && isFuture(start)) {
+    return { text: 'Coming soon', color: 'marble' };
+  } else if (!isSameDay(today, end) && isPast(end)) {
+    return { text: 'Past', color: 'marble' };
+  } else if (closesToday || closesInSevenDays) {
+    return { text: 'Final week', color: 'orange' };
+  } else {
+    return { text: 'Now on', color: 'green' };
+  }
+}
 
 const StatusIndicator: FC<Props> = ({ start, end, statusOverride }: Props) => {
   const { color, text } = statusOverride
